refactor(do): lift the this-free admin readers out of shard-do.ts
Thirteen private members that took the `SqlExec` they needed as an argument and touched no instance
state — free functions living as methods on an 11k-line class. They move to `src/admin-readers.ts`
and the call sites lose their `this.`.

Behaviour-preserving by construction: every one was `private`, and the class member list in
`api-snapshots/do.api.md` records only the 133 protected/public entries, so nothing in the frozen
surface moved and `api:check` is unchanged. No test names any of them — the 3,367-line admin suite
drives all of them through `fetch` — and all 622 pass.

Two more came along for free: `readAdminStorageReferences` and `readAdminStorageOrphans` needed only
`storageColumns()`, which takes no arguments, so its RESULT is passed rather than the hook. And
`readAdminDurableSignal` became `this`-free once its callees moved, which ESLint's
`class-methods-use-this` pointed out.

Stops there deliberately. The remaining readers reach hooks that take a table argument
(`tableRefs`, `tableColumns`, `tableIndexes`), so extracting them means passing a callback bag
rather than a value — the trade an earlier split of this file measured and rejected, and it buys
only line count. 11,273 to 11,012 is 2.3%: worth doing because it is nearly free, not because it
fixes the size. What dominates this file is the shape/subscription cluster and 36 codegen-emitted
`protected override` hooks, and neither can move.
